feat(build): add cargo deny configuration
This will allow us to catch a lot of dependency-based issues ahead of
time, like vulnerabilities, maintenance, duplication, sources. For
example, this will allow us to catch dependencies that should be using
the global level dependency, but instead uses its own and (potentially)
creates version conflicts and duplication of dependencies.